requirements.txt
# Python Requirements for Multi-Sensor LST/SST Pipeline
# Install with: pip install -r requirements.txt

# Core dependencies
earthengine-api>=0.1.300
pyyaml>=6.0
requests>=2.28.0

# Geospatial processing
GDAL>=3.4.0
rasterio>=1.3.0
rioxarray>=0.13.0

# Data processing
xarray>=2022.3.0
pandas>=1.4.0
numpy>=1.22.0
dask>=2022.1.0
netCDF4>=1.6.0

# Optional (for better performance)
bottleneck>=1.3.0
